Added `disqus:script` and enhancements
Added `{{ disqus:script }}` which only outputs the script of Disqus in
case you want to include it on a page for generating comment counts but
not actual comments.

I also added a check to see if the script has been output on the page
before so we do not output twice accidentally.

Also defaulted the ID and URL to the current `uri_string()` and current
URL for times when you are using this plugin outside of the Pages
module. This will ensure correct comment threads for all pages.